requirements.txt
sentence-transformers
scikit-learn
azure-functions
